Split up builder.rs
