+++ Cargo.toml
@@ -38,6 +38,10 @@ tar = "0.4"
 tokio = "1.0"
 url = "2.1"
 
+# XXX: This is a temporary dependency for the reexport! macro in lib.rs. Remove
+# me before 0.9.0 is released.
+paste = "1.0"
+
 [dev-dependencies]
 env_logger = "0.8"
 # Required for examples to run
